Fix: Compiler warning C4244 'argument': conversion from 'WPARAM' to 'const int32_t', possible loss of data

+++ SDF/src/Core/Window.cpp
@@ -143,11 +143,11 @@ namespace sdf {
 
     }
     void Window::onMouseInteract(WPARAM wParam, LPARAM lParam) {
-
+        
         // From MSDN: https://docs.microsoft.com/en-us/windows/win32/inputdev/wm-lbuttondown#remarks
         // To detect that the ALT key was pressed, check whether GetKeyState with VK_MENU < 0.
         // Note, this must not be GetAsyncKeyState.
-        
+
         uint32_t posX = GET_X_LPARAM(lParam);
         uint32_t posY = GET_Y_LPARAM(lParam);
 
@@ -193,7 +193,7 @@ namespace sdf {
 
         }
 
-        Input::Key key = Input::toKey(wParam);
+        Input::Key key = Input::toKey((int32_t)wParam);
         uint8_t mask = ((uint8_t)(GetKeyState(VK_CONTROL) < 0) << 0) |
                        ((uint8_t)(GetKeyState(VK_SHIFT) < 0)   << 1) |
                        ((uint8_t)(GetKeyState(VK_MENU) < 0)    << 2);
